+++ modules/runtime/include/containers/detail/sptr.hpp
@@ -125,7 +125,8 @@ skr::SPtrHelper<T, EmbedRC>::SPtrHelper(SPtrHelper<U, EmbedRC>&& lp, typename st
 template <typename T, bool EmbedRC>
 skr::SPtrHelper<T, EmbedRC>::~SPtrHelper() SKR_NOEXCEPT
 {
-    if (this->p) release();
+    if (this->p) 
+        release();
 }
 
 // methods
