+++ modules/runtime/include/containers/detail/sptr_base.hpp
@@ -190,7 +190,10 @@ skr::SPtrBase<T, EmbedRC>::SPtrBase(const SPtrBase<U, EmbedRC>& lp, T* pValue) S
     }
     else
     {
-        p->add_refcount();
+        if (auto object = sobject_cast<SInterface*>(p))
+        {
+            object->add_refcount();
+        }
     }
 }
 
@@ -208,8 +211,10 @@ skr::SPtrBase<T, EmbedRC>::SPtrBase(const SPtrBase<U, EmbedRC>& lp, typename std
     }
     else
     {
-        auto object = sobject_cast<SInterface*>(p);
-        object->add_refcount();
+        if (auto object = sobject_cast<SInterface*>(p))
+        {
+            object->add_refcount();
+        }
     }
 }
