+++ modules/runtime/src/io/build.io.cpp
@@ -1,2 +1,5 @@
-#include "io_future.cpp"
 #include "io.cpp"
+
+#include "ram_io.cpp"
+#include "vfs_ram_reader.cpp"
+// #include "dstorage_ram_reader.cpp"
