+++ modules/runtime/src/io/ram/ram_batch.hpp
@@ -1,0 +1,21 @@
+#pragma once
+#include "../io_batch.hpp"
+#include "ram_request.hpp"
+
+namespace skr {
+namespace io {
+
+struct RAMIOBatch : public IOBatchBase
+{
+    RAMIOBatch(ISmartPool<IIOBatch>* pool, uint64_t seq, uint64_t n)
+        : IOBatchBase(pool, seq)
+    {
+        reserve(n);
+    }
+
+    IOResultId add_request(IORequest request, skr_io_future_t* future) SKR_NOEXCEPT;
+};
+using RAMBatchPtr = skr::SObjectPtr<RAMIOBatch>;
+
+} // namespace io
+} // namespace skr
