+++ modules/runtime/src/io/ram/ram_buffer.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "pool.hpp"
+#include "../pool.hpp"
 
 #include "tracy/Tracy.hpp"
 
@@ -14,14 +14,9 @@ public:
     uint8_t* get_data() const SKR_NOEXCEPT { return bytes; }
     uint64_t get_size() const SKR_NOEXCEPT { return size; }
 
-protected:
     void allocate_buffer(uint64_t n) SKR_NOEXCEPT;
     void free_buffer() SKR_NOEXCEPT;
 
-    friend struct RAMIORequest;
-    friend struct RAMService;
-    friend struct RAMServiceImpl;
-    friend struct RAMBufferFuture;
     uint8_t* bytes = nullptr;
     uint64_t size = 0;
